Raspberry/requirements.txt
firebase-admin==2.12.0
gpiozero==1.4.1
pyserial==3.4
python-decouple==3.1
Adafruit-DHT